Moves internals into seperate folder
Resolves brave/brave-browser#4003
