+++ browser/resources/resource_ids
@@ -67,6 +67,10 @@
     "includes": [38700],
   },
   # This file is generated during the build.
+  "<(ROOT_GEN_DIR)/brave/web-ui-brave_rewards_internals/brave_rewards_internals.grd": {
+    "includes": [38750],
+  },
+  # This file is generated during the build.
   "<(ROOT_GEN_DIR)/brave/web-ui-brave_tip/brave_tip.grd": {
     "includes": [38800],
   },
