+++ components/brave_rewards/resources/internals/BUILD.gn
@@ -5,5 +5,5 @@ transpile_web_ui("internals") {
     ["brave_rewards_internals", rebase_path("brave_rewards_internals.tsx")]
   ]
 
-  resource_name = "brave_rewards_page"
+  resource_name = "brave_rewards_internals"
 }
