+++ components/brave_rewards/resources/internals/actions/rewards_internals_actions.ts
