+++ components/brave_rewards/resources/internals/brave_rewards_internals.tsx
