+++ components/brave_rewards/resources/internals/components/keyInfoSeed.tsx
@@ -5,7 +5,7 @@
 import * as React from 'react'
 
 // Utils
-import { getLocale } from '../../../../../common/locale'
+import { getLocale } from '../../../../common/locale'
 
 interface Props {
   isKeyInfoSeedValid: boolean
