+++ components/brave_rewards/resources/internals/components/walletPaymentId.tsx
