+++ components/brave_rewards/resources/internals/constants/rewards_internals_types.ts
